pokmon_battle/requirements.txt
aiosmtpd==1.4.5
amqp==5.2.0
asgiref==3.8.1
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
boto3==1.34.96
botocore==1.34.96
celery==5.4.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.1
Django==5.0.4
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.0
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
s3transfer==0.10.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
